Create a new task implementation for remote delivery publishing.
- Created a new task for remote delivery publishing with more precise name and method names
- Deprecated old task
