+++ model/publishing/delivery/RemotePublishingService.php
@@ -22,6 +22,7 @@ declare(strict_types=1);
 namespace oat\taoPublishing\model\publishing\delivery;
 
 use oat\taoPublishing\model\PlatformService;
+use oat\taoPublishing\model\publishing\delivery\tasks\RemoteDeliveryPublishingTask;
 use oat\taoPublishing\model\publishing\exception\PublishingInvalidArgumentException;
 use Throwable;
 use core_kernel_classes_Resource;
@@ -30,7 +31,6 @@ use oat\oatbox\service\ConfigurableService;
 use oat\tao\model\taskQueue\QueueDispatcherInterface;
 use oat\tao\model\taskQueue\Task\CallbackTaskInterface;
 use oat\taoDeliveryRdf\model\DeliveryAssemblyService;
-use oat\taoPublishing\model\publishing\delivery\tasks\DeployTestEnvironments;
 use oat\taoPublishing\model\publishing\exception\PublishingFailedException;
 
 class RemotePublishingService extends ConfigurableService
@@ -121,7 +121,7 @@ class RemotePublishingService extends ConfigurableService
             $environment->getLabel()
         );
 
-        return $this->queueDispatcher->createTask(new DeployTestEnvironments(), $params, $message);
+        return $this->queueDispatcher->createTask(new RemoteDeliveryPublishingTask(), $params, $message);
     }
 
     private function validateRemoteEnvironment(core_kernel_classes_Resource $environment): void
